all: Move EntityType to graph::schema
