+++ graph/src/schema/ast.rs
@@ -406,11 +406,10 @@ pub fn is_list(field_type: &s::Type) -> bool {
 #[test]
 fn entity_validation() {
     use crate::components::store::EntityKey;
-    use crate::components::store::EntityType;
     use crate::data::store;
     use crate::entity;
     use crate::prelude::{DeploymentHash, Entity};
-    use crate::schema::InputSchema;
+    use crate::schema::{EntityType, InputSchema};
 
     const DOCUMENT: &str = "
     enum Color { red, yellow, blue }
